% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby.R
\name{nearby}
\alias{nearby}
\title{Nearby Function}
\usage{
nearby(points, polys, column)
}
\arguments{
\item{points}{an sp SpatialPointsDataFrame}

\item{polys}{an sp SpatialPolygonsDataFrame}

\item{column}{a column in SpatialpointsDataFrame attribute table}
}
\description{
Function to find nearest polygons to a set of spatial points and return identifying column from polygons as a column in points attribute table.
}
\examples{
nearby()
}
\keyword{gDistance}
\keyword{proximity}
\keyword{spatial}
